% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipps.r
\name{bipps}
\alias{bipps}
\title{Bayesian Meshed GP regression model fit via Markov chain Monte Carlo}
\usage{
bipps(
  y,
  x,
  coords,
  k = NULL,
  family = "poisson",
  axis_partition = NULL,
  block_size = 30,
  n_samples = 1000,
  n_burnin = 100,
  n_thin = 1,
  n_threads = 4,
  verbose = 0,
  settings = list(adapting = TRUE, ps = TRUE, saving = TRUE, low_mem = FALSE, hmc = 0),
  prior = list(beta = NULL, tausq = NULL, sigmasq = NULL, phi = NULL, a = NULL, nu =
    NULL, toplim = NULL, btmlim = NULL, set_unif_bounds = NULL),
  starting = list(beta = NULL, tausq = NULL, theta = NULL, lambda = NULL, v = NULL, a =
    NULL, nu = NULL, mcmcsd = 0.05, mcmc_startfrom = 0),
  debug = list(sample_beta = TRUE, sample_tausq = TRUE, sample_theta = TRUE, sample_w =
    TRUE, sample_lambda = TRUE, verbose = FALSE, debug = FALSE),
  indpart = FALSE,
  just_preprocess = FALSE
)
}
\arguments{
\item{y}{vector of response variable}

\item{x}{matrix of covariates}

\item{coords}{matrix of spatial coordinates}

\item{k}{number of latent processes}

\item{family}{family of the response variable}

\item{axis_partition}{number of blocks in each axis}

\item{block_size}{number of observations in each block}

\item{n_samples}{number of MCMC samples}

\item{n_burnin}{number of MCMC burn-in samples}

\item{n_thin}{thinning rate for MCMC samples}

\item{n_threads}{number of threads to use}

\item{verbose}{verbosity level, integer, >= 0}

\item{settings}{list of settings for the MCMC algorithm}

\item{prior}{list of prior distribution parameters}

\item{starting}{list of starting values for MCMC}

\item{debug}{list of debugging options}

\item{indpart}{logical, whether to use independent partitioning}

\item{just_preprocess}{logical, whether to only preprocess the data}
}
\value{
list of MCMC samples
}
\description{
Bayesian Meshed GP regression model fit via Markov chain Monte Carlo
}
