% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{create_y_list}
\alias{create_y_list}
\title{Create y_list and coords objects for multi_bipps}
\usage{
create_y_list(x, y, types, image_ids, nx, ny)
}
\arguments{
\item{types}{character vector of types}

\item{image_ids}{character vector of image IDs}

\item{nx}{integer, number of pixels on x axis}

\item{ny}{integer, number of pixels on y axis}

\item{X}{vector of x coordinates}

\item{Y}{vector of y coordinates}
}
\value{
list containg y_list and coords
}
\description{
Create y_list and coords objects for multi_bipps
}
