% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{cross_cor}
\alias{cross_cor}
\title{Get cross-correlation matrix rvar at length-scale h}
\usage{
cross_cor(out, h)
}
\arguments{
\item{out}{output from multi_bipps}

\item{h}{length scale (in [0,1])}
}
\value{
cross-correlation rvar object
}
\description{
Get cross-correlation matrix rvar at length-scale h
}
