% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{cross_covar}
\alias{cross_covar}
\title{Get cross-covariance matrix rvar at length-scale h}
\usage{
cross_covar(out, h)
}
\arguments{
\item{out}{output from multi_bipps}

\item{h}{length scale (in [0,1])}
}
\value{
cross-covariance rvar object
}
\description{
Get cross-covariance matrix rvar at length-scale h
}
