% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{get_rvars}
\alias{get_rvars}
\title{Get posterior::rvar object from output of multi_bipps}
\usage{
get_rvars(out, variable)
}
\arguments{
\item{out}{multi_bipps output}

\item{variable}{string of variable to extract}
}
\value{
posterio::rvar of variable
}
\description{
Get posterior::rvar object from output of multi_bipps
}
