% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_bipps_v2.R
\name{multi_bipps_v2}
\alias{multi_bipps_v2}
\title{Runs main multi_bipps MCMC sampling}
\usage{
multi_bipps_v2(
  x,
  y,
  types,
  image_ids,
  covariates = NULL,
  k = NULL,
  nx = 20,
  ny = 20,
  n_samples = 1000,
  n_burnin = 100,
  n_thin = 1,
  n_threads = 4,
  verbose = 0,
  settings = list(adapting = TRUE, saving = TRUE, low_mem = FALSE),
  prior = list(beta = NULL, tausq = NULL, sigmasq = NULL, phi = NULL, a = NULL, nu =
    NULL, toplim = NULL, btmlim = NULL, set_unif_bounds = NULL),
  starting = list(beta = NULL, tausq = NULL, theta = NULL, lambda = NULL, v = NULL, a =
    NULL, nu = NULL, mcmcsd = 0.05, mcmc_startfrom = 0),
  debug = list(sample_beta = TRUE, sample_tausq = TRUE, sample_theta = TRUE, sample_w =
    TRUE, sample_lambda = TRUE, verbose = FALSE, debug = FALSE),
  indpart = FALSE,
  just_preprocess = FALSE
)
}
\arguments{
\item{x}{x coordinate of each cell}

\item{y}{y coordinate of each cell}

\item{types}{vector of cell types}

\item{image_ids}{vector of image that each cell comes from}

\item{k}{number of latent processes}

\item{n_samples}{number of MCMC samples}

\item{n_burnin}{number of burn in samples}

\item{n_thin}{thinning rate}

\item{n_threads}{number of threads to use}

\item{verbose}{integer of verbosity level, 0 for none}

\item{settings}{list of settings for the MCMC}

\item{prior}{list of prior parameters}

\item{starting}{list of starting values}

\item{debug}{list of debug options}

\item{indpart}{whether to use independent partitioning}

\item{just_preprocess}{whether to just preprocess the data}

\item{family}{family of the response variables}

\item{axis_partition}{number of blocks per axis}

\item{block_size}{number of observations per block}
}
\value{
list of MCMC samples
}
\description{
Runs main multi_bipps MCMC sampling
}
