% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocess.R
\name{pixellate_grid}
\alias{pixellate_grid}
\title{Pixellate a grid with spatial data and count points of each type}
\usage{
pixellate_grid(x, y, types, image_ids, nx, ny)
}
\arguments{
\item{x}{A numeric vector of x-coordinates of the points.}

\item{y}{A numeric vector of y-coordinates of the points.}

\item{types}{A character or factor vector indicating the type of each point.}

\item{image_ids}{A vector of identifiers indicating which image each point belongs to.}

\item{nx}{An integer specifying the number of pixels in the x-direction.}

\item{ny}{An integer specifying the number of pixels in the y-direction.}
}
\value{
A list with two elements:
\item{y_list}{A named list of matrices where each matrix corresponds to a unique image ID,
with rows representing pixels and columns representing counts of each point type.}
\item{coords}{A data frame containing the coordinates of each pixel center.}
}
\description{
This function takes spatial coordinates of points, scales them to fit within a
grid defined by the dimensions \verb{[0, R1] x [0, R2]}, (with R1 or R2 == 1 and
the other < 1) and pixellates the grid. It
counts the number of points of each type in each pixel, considering separate images
when image IDs are provided.
}
