% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_y_list}
\alias{plot_y_list}
\title{Plot each image in y_list, faceted by types}
\usage{
plot_y_list(y_list, coords)
}
\arguments{
\item{y_list}{y_list from create_y_list}

\item{coords}{coords from create_y_list}
}
\value{
list of ggplot2 plots
}
\description{
Plot each image in y_list, faceted by types
}
