% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.bipps.r
\name{predict.bipps}
\alias{predict.bipps}
\title{Predict method for bipps objects}
\usage{
\method{predict}{bipps}(object, newx, newcoords, n_threads = 4, verbose = FALSE, ...)
}
\arguments{
\item{object}{bipps object}

\item{newx}{matrix of new locations to predict at}

\item{newcoords}{data.frame of new locations to predict at}

\item{n_threads}{number of threads to use}

\item{verbose}{logical, print progress}

\item{...}{additional arguments}
}
\value{
list of predictions
}
\description{
Predict method for bipps objects
}
