% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{cross_cor}
\alias{cross_cor}
\title{#' Get coherence matrix rvar at length-scale h
#'
#' Calculates the coherence, e.g.,
#' \deqn{\rho_{ij}(h) = \frac{C_{ij}(h)}{\sqrt{C_{ii}(h)C_{jj}(h)}}}
#' with \eqn{C_{ij}(h)} the cross-covariance as calculated in \link{cross_covar}.}
\usage{
cross_cor(out, h, cov0, ...)
}
\arguments{
\item{out}{list of outputs from multi_bipps (one entry per chain)
#' @param h length scale (in \link{0,1})
#'
#' @return cross-correlation rvar object
#' @export
coherence <- function(out,h) {
covs <- cross_covar(out,h)
cov2cor_rfun(covs)
}}
}
\description{
#' Get coherence matrix rvar at length-scale h
#'
#' Calculates the coherence, e.g.,
#' \deqn{\rho_{ij}(h) = \frac{C_{ij}(h)}{\sqrt{C_{ii}(h)C_{jj}(h)}}}
#' with \eqn{C_{ij}(h)} the cross-covariance as calculated in \link{cross_covar}.
}
