% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{cross_covar}
\alias{cross_covar}
\title{Get cross-covariance matrix rvar at length-scale h}
\usage{
cross_covar(out, h, ...)
}
\arguments{
\item{out}{list of outputs from multi_bipps (one entry per chain)}

\item{h}{length scale (in \link{0,1})}
}
\value{
cross-covariance rvar object
}
\description{
Calculates the cross-covariance, e.g.,
\deqn{C{ij}(h) = \Lambda W \Lambda'}
with \eqn{\Lambda} estimated from MCMC and \eqn{W} the diagonal matrix with elements
\eqn{\exp{(-\phi_jh})} for \eqn{j=1,\ldots,k} on the diagonal.
}
