% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{get_rvars}
\alias{get_rvars}
\title{Get posterior::rvar object from output of multi_bipps}
\usage{
get_rvars(out_list, variable, thin = 1)
}
\arguments{
\item{out_list}{list of outputs from multi_bipps (one entry per chain)}

\item{variable}{string of variable to extract}

\item{thin}{integer, thinning interval}
}
\value{
posterio::rvar of variable
}
\description{
Get posterior::rvar object from output of multi_bipps
}
